Organize code into different views and permissions
